requirements.txt
streamlit
PyMuPDF
python-docx
sentence-transformers
openai
rapidfuzz
pandas
numpy
matplotlib
